Release emuStudio 0.41

+++ plugins/cpu/brainduck-cpu/src/main/java/net/emustudio/plugins/cpu/brainduck/BrainCPUContext.java
@@ -1,7 +1,7 @@
 /*
  * This file is part of emuStudio.
  *
- * Copyright (C) 2006-2020  Peter Jakubčo
+ * Copyright (C) 2006-2023  Peter Jakubčo
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -24,7 +24,7 @@ import net.emustudio.emulib.plugins.device.DeviceContext;
 
 @PluginContext
 public interface BrainCPUContext extends CPUContext {
-    void attachDevice(DeviceContext<Short> device);
+    void attachDevice(DeviceContext<Byte> device);
 
     void detachDevice();
 }
